Read from JSON File
Read the lectures from a local json file instead of firebase firestore.
This allows to read all lectures at once instead of fighting with pagination.

Add a layer of abstraction of the environment basics of angular in order
to inject a single service that reads from the environment files and
exposes the information