Introduce BlockDto
Adapt the model change from the backend which now carries multiple
blocks per lecture. Map the blocks back to simple strings in the Lecture
model to keep presentation easy as of now.

For later, the blocks offer flexibility if further data or more
blocks per lecture need to be displayed